test: separate entity operations from suspension gate behavior
SuspensionGateTest should explain generic buffering and pass-through behavior, not the details of each entity operation. Entity operation behavior belongs with the operation it exercises, while the generic suspension gate test should cover only gate routing. Add dedicated job, process instance, and message suspension gate test classes without dropping the existing coverage.
